% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.GenCPM.R
\name{cox.GenCPM}
\alias{cox.GenCPM}
\title{Fit a Cox Model for Survival Outcome using Connectome-based Predictive Modeling}
\usage{
cox.GenCPM(
  connectome,
  x = NULL,
  time,
  status,
  cv = "leave-one-out",
  k = dim(connectome)[3],
  thresh = 0.01,
  edge = "separate",
  seed = 1220
)
}
\arguments{
\item{connectome}{an array indicating the connectivity between M1 edges and over N subjects. The dimension must be \code{M1*M1*N}.}

\item{x}{non-image covariates matrix of \code{n (obs)* p (vars)}.}

\item{time}{the follow-up time for all individuals.}

\item{status}{the status indicator, normally 0=alive and 1=event.}

\item{cv}{a character indicating the method of cross-validation. The default method is "leave-one-out" cross-validation.}

\item{k}{a parameter used to set the number of folds for k-fold cross-validation.}

\item{thresh}{the value of the threshold for selecting significantly related edges. The default value is .01.}

\item{edge}{a character indicating the model is fitted with either positive and negative edges respectively or combined edges together. The default is "separate".}

\item{seed}{the value used to set seed for random sampling in the process of cross-validation. The default value is 1220.}
}
\value{
A list contains positive edges, negative edges, predicted survival, and actual survival outcome.
}
\description{
Fit a Cox Model for Survival Outcome using Connectome-based Predictive Modeling
}
