% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.GenCPM.R
\name{heatmap.GenCPM}
\alias{heatmap.GenCPM}
\title{Generate heatmaps for the GenCPM object based on 10-node network label in Shen268 atlas.}
\usage{
heatmap.GenCPM(cpm, foldThreshold = 0.5)
}
\arguments{
\item{cpm}{Returned GenCPM object from \code{.GenCPM} or \code{.regularized.GenCPM} functions.}

\item{foldThreshold}{the edges selected for over this many folds will be plotted. If set to .5, the edges selected at least half of the time are plotted.}
}
\description{
Generate heatmaps for the GenCPM object based on 10-node network label in Shen268 atlas.
}
