% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.regularized.GenCPM.R
\name{logit.regularized.GenCPM}
\alias{logit.regularized.GenCPM}
\title{Penalized Version of Logistic Model Prediction}
\usage{
logit.regularized.GenCPM(
  connectome,
  behavior,
  x,
  cv = "leave-one-out",
  k = dim(connectome)[3],
  thresh = 0.01,
  edge = "separate",
  type = "lasso",
  lambda = NULL,
  alpha = NULL,
  seed = 1220
)
}
\arguments{
\item{connectome}{a array indicating the connectivity between M edges and over N subjects. The dimension should be M\emph{M}N.}

\item{behavior}{a vector containing the behavior measure for all subjects.}

\item{x}{a data frame containing the non-image variables in the model.}

\item{cv}{a character indicating the method of cross-validation. The default method is "leave-one-out" cross-validation.}

\item{k}{a parameter used to set the number of folds for k-fold cross-validation.}

\item{thresh}{the value of the threshold for selecting significantly related edges. The default value is .01.}

\item{edge}{a character indicating the model is fitted with either positive and negative edges respectively or combined edges together. The default is "separate".}

\item{type}{type of penalty. The default is lasso.}

\item{lambda}{the value of penalty of LASSO regression.}

\item{alpha}{the alpha for elastic net penalty.}

\item{seed}{the value used to set seed for random sampling in the process of cross-validation. The default value is 1220.}
}
\value{
A list contains positive edges, negative edges, predicted behavior, actual behavior, and the optimal lambda.
}
\description{
Penalized Version of Logistic Model Prediction
}
