% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.GenCPM.R
\name{train.GenCPM}
\alias{train.GenCPM}
\title{Model Training using Connectome-based Predictive Modeling}
\usage{
train.GenCPM(
  train_array,
  train_behav = NULL,
  train_x = NULL,
  fit = "linear",
  thresh = 0.01,
  edge = "separate"
)
}
\arguments{
\item{train_array}{a array indicating the connectivity between M edges and over N subjects. The dimension should be \code{M*M*N}.}

\item{train_behav}{a vector containing the behavior measure for all subjects.}

\item{train_x}{a data frame containing the non-image variables in the model.}

\item{fit}{the method to be used in fitting the model. The default method is "linear".}

\item{thresh}{the value of the threshold for selecting significantly related edges. The default value is .01.}

\item{edge}{a character indicating the model is fitted with either positive and negative edges respectively or combined edges together. The default is "separate".}
}
\value{
A list contains positive edges, negative edges, the model based on positive edges and negative edges separately, or the model based on combined edges.
}
\description{
Model Training using Connectome-based Predictive Modeling
}
